-Added a check-out which displays a receipt at the end with the needed information

+++ DataModels/SeatsModel.cs
@@ -1,12 +1,12 @@
 public class SeatsModel
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string RowNumber { get; set; }
     public int SeatNumber { get; set; }
     public int SeatTypeId { get; set; }
     public double Price { get; set; }
 
-    public SeatsModel(int id, string rowNumber, int seatNumber, int seatTypeId, double price)
+    public SeatsModel(long id, string rowNumber, int seatNumber, int seatTypeId, double price)
     {
         Id = id;
         RowNumber = rowNumber;
